Rebuilt Facebook 3.1.1 binding based on MonoTouch Bindings project.
The MonoTouch Bindings project eventually updated to support 3.1.1 and it
preserves access to the old SDKs (still needed for some things) which the other
Facebook binding project I had switched to does not.

I'm not updating the version number here because in theory I'm the only guy
using this, so I know that I need to blow away my local Maven cache to get the
updated 3.1.1 artifact. Caveat coder.